REFACT: Moved EurekaMetadataParser and corresponding config. classes to core module
