+++ api-catalog-services/src/main/java/com/ca/mfaas/apicatalog/services/cached/CachedProductFamilyService.java
@@ -9,7 +9,7 @@
  */
 package com.ca.mfaas.apicatalog.services.cached;
 
-import com.ca.mfaas.apicatalog.metadata.EurekaMetadataParser;
+import com.ca.mfaas.eurekaservice.client.util.EurekaMetadataParser;
 import com.ca.mfaas.apicatalog.model.APIContainer;
 import com.ca.mfaas.apicatalog.model.APIService;
 import com.ca.mfaas.apicatalog.model.SemanticVersion;
