+++ api-catalog-services/src/main/java/com/ca/mfaas/apicatalog/services/status/APIDocRetrievalService.java
@@ -10,7 +10,7 @@
 package com.ca.mfaas.apicatalog.services.status;
 
 import com.ca.mfaas.apicatalog.instance.InstanceRetrievalService;
-import com.ca.mfaas.apicatalog.metadata.EurekaMetadataParser;
+import com.ca.mfaas.eurekaservice.client.util.EurekaMetadataParser;
 import com.ca.mfaas.apicatalog.services.cached.model.ApiDocInfo;
 import com.ca.mfaas.apicatalog.services.status.model.ApiDocNotFoundException;
 import com.ca.mfaas.apicatalog.swagger.SubstituteSwaggerGenerator;
