+++ api-catalog-services/src/test/java/com/ca/mfaas/apicatalog/swagger/SubstituteSwaggerGeneratorTest.java
@@ -9,7 +9,7 @@
  */
 package com.ca.mfaas.apicatalog.swagger;
 
-import com.ca.mfaas.apicatalog.metadata.EurekaMetadataParser;
+import com.ca.mfaas.eurekaservice.client.util.EurekaMetadataParser;
 import com.ca.mfaas.config.ApiInfo;
 import com.netflix.appinfo.InstanceInfo;
 import com.netflix.appinfo.InstanceInfo.PortType;
