+++ common-service-core/src/main/java/com/ca/mfaas/product/utils/UrlUtils.java
